+++ src/main/java/com/webstore/entity/ProductPrice.java
@@ -1,19 +1,10 @@
 package com.webstore.entity;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.SequenceGenerator;
-import jakarta.persistence.Table;
-import jakarta.persistence.UniqueConstraint;
-
-import lombok.EqualsAndHashCode;
+import jakarta.persistence.*;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+import java.math.BigInteger;
 
 @Data
 @Entity
@@ -34,17 +25,14 @@ public class ProductPrice extends BasicEntities {
     @Column(name = "product_price_id")
     private Integer productPriceId;
 
-
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "product_id", nullable = false)
     private Product product;
-
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "currency_id", nullable = false)
     private Currency currency;
 
     @Column(name = "price_amount", nullable = false)
-    private Long priceAmount;
-
+    private BigInteger priceAmount;
 }
